Modify overlay node combination algorithm
When fragment 'a' was referring to node 1 in the
base device tree, and fragment 'b' was referring
to a child of node 1, the earlier overlay node
combination algorithm was not combining
fragment 1 and fragment 2. This can cause the
verification algorithm to cause a false failure.

The test apply_verify_multiple_overlay.sh has been
modified to test this case.

Bug: 67779848
Test: sh apply_verify_multiple_overlay.sh
Change-Id: I7a8ff2e0fb75c0346f8d208ebddad62d0b2a5671