Revert "mailbox: msm_qmp: Remove enable_irq_wake"
This reverts commit 040706b97601ca1c2f549631889a6723b7bd29c1.

Reason for revert: Partner reverted change as a result of final technical
solution moving in a different direction.

A newer set of patches switches the drivers that had been configured to
use IRQF_NO_SUSPEND to use enable_irq_wake() instead, and the patch for
this driver is based on this change being reverted first.

Change-Id: I072280b91b100b8a8a0817700184b657023b5b58
Bug: 123377615
Bug: 131260677
Signed-off-by: Kelly Rossmoyer <krossmo@google.com>
Signed-off-by: UtsavisGreat <utsavbalar1231@gmail.com>