qcacld-3.0: Rectify max value of gEnableDynamicDTIM
As a part of the change I3cd617e52e86e2a890e35edb33b37ba3df8f4f6e
the max value of the ini gEnableDynamicDTIM was changed from 9 to 5.
This change causes regression where DUT is not entering in to DTIM9.

Restore back the max value of the ini to 9.

Change-Id: Ifea0c5199951a0dffc5cd116ec4c065f1563fe94
CRs-Fixed: 2489494