qcacmn: Correct additions to empty defrag waitlist
In dp_rx_defrag_waitlist_add function, if TAILQ is empty then
TAILQ_INSERT_HEAD should be called instead of TAILQ_INSERT_TAIL
as it correctly sets the tqe_prev pointer to the head of the queue
if it is the only element in the queue. Calling TAILQ_INSERT_TAIL here
will set the tqe_prev pointer to NULL, potentially leading to a
null pointer dereference error in the future.

Change-Id: I04d6a107650fd1f2a5c54ba2b9c8efc6dd9ce3f0
CRs-Fixed: 2502646