init: support early userspace
Some initcall stage takes long time to finish and it's
better we launch some of our userspace applications
earlier before these initcalls. This way, the userspace
applications we care after system booting can start to
work earlier and provide better user experience.
To achieve this, we find all ufs and rootfs related
dependencies and run them in order in a separate thread
and start this chain as early as possible.
And to make the change compatible, the bottom line is
this feature is disabled by default and can be enabled
optionally and normal booting procedure should not be
influenced if this feature is disabled. So we add a
new cmdline parameter called "early_userspace" for
this purpose.

Change-Id: I2207a4a28517ea92389a4cbcdc26cfee39d3e5b4
Signed-off-by: Jiacheng Zheng <jiaczhen@codeaurora.org>