drivers: rmnet_perf: Fix SSR cleanup of structs
When SSR occurs rmnet_mhi0 is torn down and the check perf makes
on the netdev to make sure that rmnet_mhi0 is registered with
rmnet is not valid if rmnet has processed the unregister event
before rmnet_perf.This change adds priority to rmnet_perf so that
it processes events before the main driver.

Previously rmnet_perf was not cleaning up as a result of that and
had a memleak when it would come back up. This fixes it so that perf
cleans up all of the remaining flows in rmnet_perf hashatable.

Change-Id: Ifd10a2ca586c33bf0d4fe479d54a41f8f46b720b
Acked-by: Raul Martinez <mraul@qti.qualcomm.com>
Signed-off-by: Subash Abhinov Kasiviswanathan <subashab@codeaurora.org