init: define early init functions for rootfs mounting
Here we define init functions for our first subsystem which
initializes flash controller, filesystem, and starts our
early userspace applications.
And ufs scan runs in sync mode if early feature is enabled
so it runs in the thread for rootfs mounting and we do not
need additional complicated code to do the sync job.

Change-Id: I9001e5076a4b0844f78f2a75f91a6b01f3f1bc6d
Signed-off-by: Jiacheng Zheng <jiaczhen@codeaurora.org>