init: define early devices needed by all subsystems
Devices such as clocks, regulators, bus components, iommu,
etc. are defined here. They are commonly needed by all
subsystems so we populate them at the very beginning of
our early platform driver. And different soc and boards
may differ in this list, so device tree is a good place
to define this list.

Change-Id: Ia616f2e7bf289d66f2bb5de026726e8a5a387df7
Signed-off-by: Jiacheng Zheng <jiaczhen@codeaurora.org>