msm: kgsl: Skip crashdumper if MMU is not attached
Crashdumper requires its context bank initialized and a valid pagetable
attached. So skip crasdumper if this is not the case during recovery.

This check is required because GMU might leave GX HS ON before we
initialize MMU during GPU wake up which in turn will trigger crashdumper
while collecting snapshot.

Change-Id: I6b3852718139cd2a2241b526755bfc797ac1f76a
Signed-off-by: Akhil P Oommen <akhilpo@codeaurora.org>
Signed-off-by: Pankaj Gupta <gpankaj@codeaurora.org>
Signed-off-by: UtsavBalar1231 <utsavbalar1231@gmail.com>