qcacmn: Place the max BW support above the ch params
In the structure scan_chan_list_params the max bw support
is present below the ch params which is used as an pointer
to a memory, and the max bw support can overwrite the
memory which is pointed by ch params.

Fix is to place the BW support above the ch params array.

Change-Id: I997c39b4b87e05ff695d114ffeb4cec31397fbdf
CRs-Fixed: 2564456