mhi: core: Check for queued request instead of available request
commit 741a49c45b3f2dac ("mhi: core: Handle RSC minimum credit
requirement") checking available space in transfer ring to
queue against the minimum RSC credit. This causes data stall
if transfer ring is almost full and OOB event is received. Due to
that minimum credit requirement does not match and Host never
rings the DB and device does not post event for host to process.
Fix this issue by checking queued request against minimum credit
which allows host to ring DB when ring is almost full and OOB is
received.

Change-Id: I3ab88aff97d78db003c49e888802e77878c95fa2
Signed-off-by: Hemant Kumar <hemantk@codeaurora.org>
Signed-off-by: UtsavisGreat <utsavbalar1231@gmail.com>