drivers: shs: protect mmap file operations using shs ep lock
This adds synchronization for the mmap file operation so that an mmap
will only occur after a succesful open and will not occur after a release.
This is done by using a fault handler for each shared memory file and
checking the respective global variable inside of the ep lock.

Change-Id: I955dfc8e745b1e275328721e9934ef8c020d8fa3
Signed-off-by: Subash Abhinov Kasiviswanathan <subashab@codeaurora.org>