clk: qcom: mdss: DSI and DP PLL changes for SDM660
This is a snapshot of mdss-dp-14nm files as of commit
'21e7c25c87a4c270 ("clk: qcom: mdss: update the clk_ops
for dp_vco_divided_clk_src_mux").
In addition, the parent name for dsi_vco_14nm clk was
changed from 'bi_tcxo' to 'xo_board' to be compatible
with the clk driver used for SDM660.

Change-Id: Iaae770c8ef71f5304790254042ac7e8de504d451
Signed-off-by: Nirmal Abraham <nabrah@codeaurora.org>
Signed-off-by: Raghavendra Ambadas <rambad@codeaurora.org>