coresight: replicator: Add CoreSight Replicator driver
This driver manages CoreSight Replicator device, which
resides on the AMBA bus. Replicator has been made programmable to
allow software to turn of the replicator branch to sink that is not
being used. This avoids trace traffic to the unused/non-current sink
from causing back pressure that results in overflows at the source.
This is snapshot of the coresight-replicator-qcom driver as of
msm-4.4 'commit e4c6a37f8770697d4 ("msm: vidc: avoid OOB
write while accessing memory").

Change-Id: Ibb6fa20aa1820eb8b08523200b7aad49a4355bd9
Signed-off-by: Chetan C R <cchinnad@codeaurora.org>
Signed-off-by: Anusha Shakarad <ashaka@codeaurora.org>