fbdev: msm: Use dynamic allocation for SID variable.
The physical address of SID variable when passed to TZ during
SCM call for secure display was causing a out of range address
error from TZ followed by SMMU fault. The SID was a
defined as a local stack variable. Use dynamic allocation for
SID variable to avoid this address out of range error.

Change-Id: Ia1a49ccdf2b386b6d600676ecf5a2d4a53b993b9
Signed-off-by: Nirmal Abraham <nabrah@codeaurora.org>
Signed-off-by: Rajshekar Eashwarappa <reashw@codeaurora.org>