mm/free_area_init_core: skip memmap_init for holes
For every page in the memory holes, we check and find it's
not in any valid memory block and return without doing
anything. This is not good and wastes a lot time especially
when the hole is very big. So we should skip pages in holes
and call memmap_init only for pages in valid memory blocks.

Change-Id: Ida7b0631c064f8493ea1c5e4d4d24d9b88288da3
Signed-off-by: Jiacheng Zheng <jiaczhen@codeaurora.org>