qcacld-3.0: Fix Peer leak if vdev is logically deleted by SSR
If Vdev delete is in progress and SSR is initiated the HDD wait is
aborted and Vdev is logically deleted.

In wma_remove_objmgr_peer driver tries to get the vdev ref which
fails as vdev is logically deleted and thus peer is not deleted.

Fix is to use wma->pdev instead of getting pdev from vdev and then
find peer on pdev.

Change-Id: Ib32e95a17c70f12add49cb8db314e311c8b2c1c1
CRs-Fixed: 2649071