ARM: dts: msm: Add a vote for CPU to GPU Config path for sdmmagpie
We should have a vote for CPU to GPU config path
when we are trying to access any GPU config registers.
This will make sure that all the clocks in the CPU to
GPU config paths are ON and we are not riding on
someone's else vote for this path.

Change-Id: Ib6f0bb0730a2b4e267b053366385179b9d8387a8
Signed-off-by: Indira Biruduraju <ibirudur@codeaurora.org>