f2fs: add symbolic link to kobject in sysfs
Added a symbolic link to directory of sysfs. It will
create a symbolic link such as "mount_0" and "mount_1" to
each f2fs mount in the order of mounting filesystem. It will
provide easy access to sysfs node even if not knowing the
specific device node name like sda19 and dm-3.

Signed-off-by: Daeho Jeong <daehojeong@google.com>
Signed-off-by: Jaegeuk Kim <jaegeuk@kernel.org>