ARM64: configs: raphael: Make Ion allocate in page-sized chunks
With this option disabled, Ion allocates memory in order-4 sized chunks
(16 pages at a time), which is quite a heavy burden on the page
allocator, and can take a while to allocate when memory is fragmented
after a lot of uptime. To lower the allocation latency in Ion, have it
only allocate in page-sized chunks.

Change-Id: I367f27ca4eeb72eee5bfb96edd8a17248bafd798
Signed-off-by: Sultan Alsawaf <sultan@kerneltoast.com>
Signed-off-by: UtsavBalar1231 <utsavbalar1231@gmail.com>