ARM: dts: msm: move certain properties to overlay device-tree
Certain behavior altering properties such as offload-chan or
lpm-notify must be moved to their respective overlay device-tree
files as those may need to be modified from time to time if the
channel usage changes based on the device. It can also cause
unexpected outcomes such as crashes on the host or device.

Change-Id: I0341b60e57e7b714cb92c50316c27faf4df94fad
Signed-off-by: Bhaumik Bhatt <bbhatt@codeaurora.org>