mhi: cntrl: qcom: enable l1ss timeout if skipping force suspend
In some controllers, there is no need to force a suspend after
mission mode entry in host. However, the L1ss timeout needs to be
enabled if skipping forced suspends. Failure to do so will result
in reliance on future clients to force a runtime suspend or we
may see suspends enabled only after a system suspend occurs
sometime later. This can cause power drain and must be solved.

Change-Id: I8c8fd5fe81e21b03e03e8d3442ee63343d918478
Signed-off-by: Bhaumik Bhatt <bbhatt@codeaurora.org>