pci: switch: qcom: add support for Port Arbitration configuration
For some usecases, it is required and beneficial to give one
port a higher priority over another. Add the support so that a
switch can define which Port Arbitration scheme to select and
the values for its Port Arbitration table.

Enable 4-bit 128 phase Port Arbitration for Diode switch.
Partition so that 7 phases are designated for port1 follow by 1
phase for port2.

Change-Id: I2f0e6775e8497bcc0cfe249ea786be0dcb9500d6
Signed-off-by: Tony Truong <truong@codeaurora.org>